[render] Split up renderer into multiple files
